move hostname to profile to set one time + add - for parsing
